web: Move create_renderer to builder
